+++ product-service/src/main/java/com/shipflow/productservice/infrastructure/messaging/config/StockRestoreHandler.java
@@ -5,9 +5,9 @@ import org.springframework.stereotype.Component;
 import com.shipflow.common.messaging.handler.AbstractSagaHandler;
 import com.shipflow.common.messaging.publisher.EventPublisher;
 import com.shipflow.productservice.application.service.ProductService;
-import com.shipflow.productservice.infrastructure.messaging.OrderCanceledEvent;
-import com.shipflow.productservice.infrastructure.messaging.StockRestoredEvent;
-import com.shipflow.productservice.infrastructure.messaging.StockRestoredFailedEvent;
+import com.shipflow.productservice.infrastructure.messaging.event.OrderCanceledEvent;
+import com.shipflow.productservice.infrastructure.messaging.event.StockRestoredEvent;
+import com.shipflow.productservice.infrastructure.messaging.event.StockRestoredFailedEvent;
 
 import lombok.RequiredArgsConstructor;
 
@@ -21,11 +21,13 @@ public class StockRestoreHandler extends AbstractSagaHandler<OrderCanceledEvent>
 	public void process(OrderCanceledEvent event) {
 		try {
 			productService.restoreStock(event.getProductId(), event.getQuantity());
-			eventPublisher.publish(new StockRestoredEvent(
-				event.getOrderId(), event.getProductId(), event.getQuantity()));
+
 		} catch (Exception e) {
 			eventPublisher.publish(new StockRestoredFailedEvent(
 				event.getOrderId(), event.getProductId(), e.getMessage()));
 		}
+
+		eventPublisher.publish(new StockRestoredEvent(
+			event.getOrderId(), event.getProductId(), event.getQuantity()));
 	}
 }
