+++ product-service/src/main/java/com/shipflow/productservice/infrastructure/messaging/event/OrderCanceledEvent.java
@@ -1,4 +1,4 @@
-package com.shipflow.productservice.infrastructure.messaging;
+package com.shipflow.productservice.infrastructure.messaging.event;
 
 import com.shipflow.common.messaging.event.SagaEvent;
 
@@ -7,15 +7,14 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-public class StockRestoredEvent extends SagaEvent {
-
-	private static final String EVENT_TYPE = "stock.restored";
+public class OrderCanceledEvent extends SagaEvent {
+	private static final String EVENT_TYPE = "order.canceled";
 
 	private String orderId;
 	private String productId;
 	private Integer quantity;
 
-	public StockRestoredEvent(String orderId, String productId, Integer quantity) {
+	public OrderCanceledEvent(String orderId, String productId, Integer quantity) {
 		super(EVENT_TYPE);
 		this.orderId = orderId;
 		this.productId = productId;
