+++ product-service/src/main/java/com/shipflow/productservice/infrastructure/messaging/event/OrderCreationStartedEvent.java
@@ -1,4 +1,4 @@
-package com.shipflow.productservice.infrastructure.messaging;
+package com.shipflow.productservice.infrastructure.messaging.event;
 
 import com.shipflow.common.messaging.event.SagaEvent;
 
@@ -7,19 +7,18 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-public class ProductStockDecreasedEvent extends SagaEvent {
+public class OrderCreationStartedEvent extends SagaEvent {
 
-	private static final String EVENT_TYPE = "product.stock.decreased";
+	private static final String EVENT_TYPE = "order.creation.started";
 
 	private String orderId;
 	private String productId;
 	private Integer quantity;
 
-	public ProductStockDecreasedEvent(String orderId, String productId, Integer quantity) {
+	public OrderCreationStartedEvent(String orderId, String productId, Integer quantity) {
 		super(EVENT_TYPE);
 		this.orderId = orderId;
 		this.productId = productId;
 		this.quantity = quantity;
-
 	}
 }
